+++ src/main/java/com/structurizr/export/mermaid/MermaidDiagram.java
@@ -1,0 +1,17 @@
+package com.structurizr.export.mermaid;
+
+import com.structurizr.export.Diagram;
+import com.structurizr.view.View;
+
+public class MermaidDiagram extends Diagram {
+
+    public MermaidDiagram(View view, String definition) {
+        super(view, definition);
+    }
+
+    @Override
+    public String getFileExtension() {
+        return "mmd";
+    }
+
+}