+++ src/main/java/com/structurizr/export/plantuml/AbstractPlantUMLExporter.java
@@ -1,6 +1,7 @@
 package com.structurizr.export.plantuml;
 
-import com.structurizr.export.AbstractDiagramExporter;
+import com.structurizr.export.DiagramExporter;
+import com.structurizr.export.Diagram;
 import com.structurizr.export.IndentingWriter;
 import com.structurizr.model.*;
 import com.structurizr.util.StringUtils;
@@ -11,12 +12,13 @@ import java.util.Map;
 
 import static java.lang.String.format;
 
-public abstract class AbstractPlantUMLExporter extends AbstractDiagramExporter {
+public abstract class AbstractPlantUMLExporter extends DiagramExporter {
 
     public static final String PLANTUML_TITLE_PROPERTY = "plantuml.title";
     public static final String PLANTUML_LEGEND_PROPERTY = "plantuml.legend";
     public static final String PLANTUML_INCLUDES_PROPERTY = "plantuml.includes";
     public static final String PLANTUML_SEQUENCE_DIAGRAMS_PROPERTY = "plantuml.sequenceDiagrams";
+    public static final String PLANTUML_ANIMATION_PROPERTY = "plantuml.animation";
 
     /**
      * <p>Set this property to <code>true</code> by calling {@link Configuration#addProperty(String, String)} in your
@@ -192,6 +194,11 @@ public abstract class AbstractPlantUMLExporter extends AbstractDiagramExporter {
     }
 
     @Override
+    protected boolean isAnimationSupported(View view) {
+        return "true".equalsIgnoreCase(view.getViewSet().getConfiguration().getProperties().getOrDefault(PLANTUML_ANIMATION_PROPERTY, "false"));
+    }
+
+    @Override
     protected void writeHeader(View view, IndentingWriter writer) {
         writer.writeLine("@startuml");
 
@@ -248,6 +255,11 @@ public abstract class AbstractPlantUMLExporter extends AbstractDiagramExporter {
     @Override
     protected void writeFooter(View view, IndentingWriter writer) {
         writer.writeLine("@enduml");
+    }
+
+    @Override
+    protected Diagram createDiagram(View view, String definition) {
+        return new PlantUMLDiagram(view, definition);
     }
 
 }